Add gradient variant for Heading
