--- apps/client/src/utils/theme.ts
@@ -1,12 +1,0 @@
-import { extendTheme, ThemeConfig } from '@chakra-ui/react';
-
-const config: ThemeConfig = {
-  useSystemColorMode: true,
-  initialColorMode: 'dark',
-};
-
-const theme = extendTheme({
-  config,
-});
-
-export default theme;
